feat: an equivalence harness, and the three disagreements it found
tests/equivalence/run.py is the single entry point for the validation
strategy: run both implementations over many seeds on the same example,
reduce each result file to one value per (scenario, year, sex, variable) by
the count-weighted mean over age bands, and compare the mean, standard
deviation and 5th, 50th and 95th percentiles of each of those series. The
thresholds are a multiplicity-corrected k-sigma rule rather than a tolerance
on a single number, with a floor set by the six significant digits the
baseline's CSV carries. The baseline's reduced output is cached keyed by the
hash of the config that produced it.

Running it against the baseline on the converted HLM_France example found
three real disagreements:

- mean_gender reported 1/count instead of 1 for every male band. `gender` is
  a declared level-0 risk factor, so it is in the mapping, but a person does
  not carry it in risk_factors — so the sum came from the explicit
  accumulation and was then divided by the head count twice.

- The calibration adjustment was clamped to each factor's configured range.
  The baseline's HLM path passes no range, and that is what makes
  adjust_to_factors_mean mean what it says: shifting a band by
  `expected - simulated_mean` lands its mean exactly on the FactorsMean
  table, and clamping moves it back off. Most visible at the young ages,
  where the expected BMI of 14 sits just above the configured bound of 13.88.
  The baseline's band means are seed-independent for this reason; ours now
  are too, which is why the trial-runs test now distinguishes runs by their
  spread rather than their means.

- The expected age-sex population was scaled by size_fraction rather than by
  the whole number of people the run actually started with, so some bands fell
  on the other side of a rounding boundary and the population differed from
  the baseline's by a person or two every year.

With those fixed: 38,260 comparisons over 20 seeds of 2010-2050, both
scenarios, and every single mean agrees — 0 of 7,652. 54 comparisons fail,
all between 1.01 and 1.36 times the allowance, 29 of them on standard
deviations, concentrated in a few late years for females.

+++ src/sim/engine.cpp
@@ -105,8 +105,14 @@ model::IntegerAgeGenderTable Engine::expected_population() const {
     const auto start_year = context_.start_time();
     const auto real_start_population =
         modules_.demographic->get_total_population_size(start_year);
-    const auto cohort_size = inputs_->settings().size_fraction *
-                             static_cast<double>(real_start_population);
+    // The cohort size is the *whole number* of people the run actually started with, not the
+    // unrounded fraction. That matters: the scale below is applied to a projected age-sex count
+    // and rounded to a person, so scaling by `cohort_size / real_population` rather than by
+    // `size_fraction` moves some bands across the rounding boundary. Using the fraction directly
+    // left the reference example's population differing from the baseline's by a person or two
+    // per year, which then showed up in every count-weighted aggregate.
+    const auto cohort_size = static_cast<double>(static_cast<std::size_t>(
+        inputs_->settings().size_fraction * static_cast<double>(real_start_population)));
 
     const auto &distribution =
         modules_.demographic->get_population_distribution(context_.time_now());
